Make it possible to add custom webhook handling behavior without overriding default behavior
